fix bf16 issue

+++ src/model/load_model.py
@@ -97,7 +97,7 @@ def enable_safe_bf16_causal_mask_patch() -> bool:
 DEFAULT_BNB_CONFIG = BitsAndBytesConfig(
     load_in_4bit=True,
     bnb_4bit_quant_type="nf4",
-    bnb_4bit_compute_dtype=torch.float16,
+    bnb_4bit_compute_dtype=torch.bfloat16,
     bnb_4bit_use_double_quant=True,
 )
